+++ packages/assets-controllers/CHANGELOG.md
@@ -9,7 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add a new DeFiPositionsController that keeps an update list of DeFi positions for EVM accounts ([#5400](https://github.com/MetaMask/core/pull/5400))
+- Add a new `DeFiPositionsController` that keeps an update list of DeFi positions for EVM accounts ([#5400](https://github.com/MetaMask/core/pull/5400))
+- Adds new exports for the following types:
+  - DeFiPositionsControllerState,
+  - DeFiPositionsControllerAllowedActions,
+  - DeFiPositionsControllerAllowedEvents,
+  - DeFiPositionsControllerActions,
+  - DeFiPositionsControllerEvents,
+  - DeFiPositionsControllerGetStateAction,
+  - DeFiPositionsControllerStateChangeEvent,
+  - DeFiPositionsControllerMessenger
+  - GroupedDeFiPositions
+- Add `@metamask/transaction-controller` as a peer dependency
 
 ## [56.0.0]